11_Projects/Knowledge_Pipeline/requirements.txt
# Knowledge Pipeline - Dependencies
yt-dlp>=2024.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0
